Revert CleanUp
